Add find_unused_true.
